// rdpms-web-ui/src/lib/api_client/model/DataSetSummaryDTO.ts
/**
 * My API v1
 * API Version v1
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import * as models from './models';

/**
 * Represents a summary of a dataset, including identifying information, timestamps, state, tags,  and metadata fields.
 */
export interface DataSetSummaryDTO {
    /**
     * Uniquely identifies the dataset. Typically server-generated. Should not be manually set by the client.
     */
    id?: string;

    /**
     * Non-unique, mandatory descriptive name of the dataset. Must be provided by the client.
     */
    name?: string;

    /**
     * List of tags associated with the dataset, used for categorization and filtering purposes.
     */
    assignedTags?: Array<models.TagDTO>;

    /**
     * UTC timestamp when the dataset was originally created.  Mandatory property, should be provided by the client.
     */
    createdStampUTC?: string;

    /**
     * UTC timestamp that indicates when the dataset was deleted.  Null if the dataset has not been deleted yet.
     */
    deletedStampUTC?: string;

    /**
     * UTC timestamp that marks the period begin of the dataset.  Only to be set by server.
     */
    beginStampUTC?: string;

    /**
     * UTC timestamp that marks the period end of the dataset.  Only to be set by server.
     */
    endStampUTC?: string;

    state?: models.DataSetStateDTO;

    /**
     * Indicates if the dataset represents time-series data.  Only to be set by server.
     */
    isTimeSeries?: boolean;

    /**
     * Flags whether the dataset has been marked as deleted.  Only to be set by server.
     */
    isDeleted?: boolean;

    /**
     * Fields, for which metadata exists.  Only to be set by server.
     */
    metadataFields?: Array<string>;

}
export namespace DataSetSummaryDTO {
}